EntryRepository: use build flavor for demo repository
